✨  feat : question 생성, 수정 시 tag 기능 test 완성 #34

+++ Server/src/main/java/sixman/stackoverflow/domain/question/controller/dto/QuestionUpdateApiRequest.java
@@ -24,6 +24,6 @@ public class QuestionUpdateApiRequest {
     private String expect;
 
     @NotNull(message = "{validation.question.tag}")
-    private List<Integer> tagIds;
+    private List<Long> tagIds;
 
 }
